+++ packages/plugin-packages/mashroom-http-proxy/src/metrics/connection_pool_metrics.ts
@@ -21,14 +21,14 @@ export const startExportPoolMetrics = (pluginContextHolder: MashroomPluginContex
             if (httpPoolStats) {
                 collectorService.gauge('mashroom_http_proxy_active_connections_total', 'Mashroom HTTP Proxy Active Connections Total').set(httpPoolStats.activeConnections);
                 collectorService.gauge('mashroom_http_proxy_idle_connections_total', 'Mashroom HTTP Proxy Idle (Free) Connections Total').set(httpPoolStats.idleConnections);
-                collectorService.gauge('mashroom_http_proxy_max_connections', 'Mashroom HTTP Proxy Max Connections').set(config.maxSockets);
+                collectorService.gauge('mashroom_http_proxy_max_connections', 'Mashroom HTTP Proxy Max Connections').set(config.maxTotalSockets);
                 collectorService.gauge('mashroom_http_proxy_waiting_requests_total', 'Mashroom HTTP Proxy Waiting Requests (in Queue)').set(httpPoolStats.waitingRequests);
             }
 
             if (httpsPoolStats) {
                 collectorService.gauge('mashroom_https_proxy_active_connections_total', 'Mashroom HTTPS Proxy Active Connections Total').set(httpsPoolStats.activeConnections);
                 collectorService.gauge('mashroom_https_proxy_idle_connections_total', 'Mashroom HTTPS Proxy Idle (Free) Connections Total').set(httpsPoolStats.idleConnections);
-                collectorService.gauge('mashroom_https_proxy_max_connections', 'Mashroom HTTPS Proxy Max Connections').set(config.maxSockets);
+                collectorService.gauge('mashroom_https_proxy_max_connections', 'Mashroom HTTPS Proxy Max Connections').set(config.maxTotalSockets);
                 collectorService.gauge('mashroom_https_proxy_waiting_requests_total', 'Mashroom HTTPS Proxy Waiting Requests (in Queue)').set(httpsPoolStats.waitingRequests);
             }
         }